Fix ES index name generation to allow multiple runs per day
The ES index name is generated by formatting current date as ddMMyyyy
which causes failure if a build is run more than once a day since there
is a check to not re-create an existing index (or delete it).

This change amends the name generation to use ddMMyyyy-HHmmss to allow
running multiple builds a day.